refactor(rust,python): "model(s)"という名のモジュールを改名する
