+++ crates/voicevox_core/src/engine/talk.rs
@@ -1,13 +1,13 @@
+mod audio_query;
 mod full_context_label;
 mod interpret_query;
 mod kana_parser;
-mod model;
 pub(crate) mod open_jtalk;
 pub(crate) mod text;
 pub(crate) mod text_analyzer;
 pub(crate) mod user_dict;
 
+pub use self::audio_query::{AccentPhrase, AudioQuery, Mora};
 pub(crate) use self::full_context_label::extract_full_context_label;
 pub(crate) use self::interpret_query::{initial_process, split_mora, DecoderFeature};
 pub(crate) use self::kana_parser::{create_kana, parse_kana, KanaParseError};
-pub use self::model::{AccentPhrase, AudioQuery, Mora};
