+++ crates/voicevox_core_python_api/python/voicevox_core/_python/_please_do_not_use.py
